Split Google Chat adapter e2e tests into inbound/outbound/round-trip files
Extract the shared fixtures (fake Pub/Sub, fake chat API, state helpers,
trpc readiness proxy) plus the env-setup boilerplate into a
_google-chat-fixtures.ts helper, and split the 660-line monolith into
three focused files that each exercise one half of the adapter.
